Correctly wrap char / gchar values.
Previously, these were intepreted like as a 'char *' argument and the wrapping method expected a string. When marshalling, the pointer was also then incorrectly cast to a byte, producing garbage values
Now, the wrapper method accepts a char value, which is converted to ASCII when marshalling.

This is necessary for g_application_add_main_option(), for example.